backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.7
anthropic==0.7.8
boto3==1.29.7
moviepy==1.0.3
pillow==10.1.0
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
celery==5.3.4
edge-tts==6.1.9
pydub==0.25.1
